requirements.txt
beautifulsoup4==4.4.1
certifi==2016.2.28
future==0.15.2
python-telegram-bot==5.0.0
requests==2.10.0
urllib3==1.16
wikipedia
